+++ src/test/java/com/artipie/composer/http/PhpComposerTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artpie.composer.http;
+package com.artipie.composer.http;
 
 import com.artipie.asto.Key;
 import com.artipie.asto.Storage;
