+++ src/test/java/com/artipie/composer/package-info.java
@@ -22,24 +22,10 @@
  * SOFTWARE.
  */
 
-package com.artpie.composer;
-
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Test;
-
 /**
- * Tests for {@link Name}.
+ * PHP Composer repository tests.
  *
  * @since 0.1
  */
-class NameTest {
 
-    @Test
-    void shouldGenerateKey() {
-        MatcherAssert.assertThat(
-            new Name("vendor/package").key().string(),
-            Matchers.is("vendor/package.json")
-        );
-    }
-}
+package com.artipie.composer;
